requirements.txt
face_recognition==1.3.0
numpy==1.22.3
opencv_python==4.5.5.64
pandas==1.4.2
xlrd<2.0
Pillow==9.1.1
streamlit==1.9.2
